LogViewForm: follow bottom of the log
Minor improvement of log viewer form: if vertical scrollbar is on the
bottom and new text is appended, scroll to bottom after appending it so
that we can still see the newest entries. If the scrollbar is not on the
bottom, preserve its position so that we can still see the same entry
(as it used to be).