Skip over the mutex field when copying t_call_info data
t_line::get_call_info() returns a copy of a (locked) t_call_info;
copying the mutex, aside from being undefined in POSIX, returns an
unexpectedly locked copy, possibly resulting in a deadlock.

This fix is pretty much a copy of the one applied to t_call_record in
38bb6b7, with an added check for self-assignment.  Like its predecessor,
it is still susceptible to deadlock, though (as unlikely as it may be).

Fixes #73