Don't require file(1) when checking for libmagic
Attempting to compile Twinkle on a system where file(1) is not available
currently produces a confusing error message about libmagic not being
found.

The only reason to require file(1) was to obtain the libmagic version,
as the MAGIC_VERSION constant was apparently only introduced in 5.13.
But since we don't need any particular version, we might as well drop
this requirement.

Using find_program() instead of find_path() avoids picking
/usr/include/file by mistake, which resulted in a (harmless) empty
version string in the CMake output.

(Thanks to https://bro-tracker.atlassian.net/browse/BIT-1096 for
providing some of this information.)