Add compile-time check for C++11 regular expressions (Closes #112)
libstdc++ 4.8 shipped with a buggy prototype for C++11 regular
expressions, which barely supports the simplest cases, and will throw an
exception at pretty much anything else.  Unfortunately, this does not
cause the build to fail (since all the symbols are properly exported),
and results instead in mysterious runtime errors.  (See issue #31 for
such an example.)