Make Readline non-blocking: Use Readline's callback interface
When Twinkle is running in CLI mode and is sent a "quit" command to its
local socket, it will currently not respond immediately, but rather wait
until the next line has been read from its standard input (issue #143).

This is due to the blocking nature of readline(), which only returns
once a complete line has been read.  Switching to Readline's "alternate"
callback interface is the first step in addressing this issue.

(As a bonus, this also fixes a bug where the line pointer returned by
readline() was not freed correctly.)