Make Readline non-blocking: Relay SIGWINCH to Readline
When using Readline's callback interface, signals are (by default) only
captured within the duration of the rl_callback_read_char() call.  It is
therefore expected of the application to capture SIGWINCH on its own,
and notify Readline of the fact.

(While it's possible to change this with rl_persistent_signal_handlers,
some signals, such as SIGHUP, will still only be *processed* during that
call, making it a somewhat unappealing solution.  This all could be
alleviated by calling rl_check_signals() periodically, but that function
was only introduced in Readline 8.0, which is far too recent.)

Note that we are using signal(2) and not sigaction(2), depite the
various warnings that come with it, mostly because it's what is already
present in the codebase.