G.729B: Add support for dealing with G.729B SID frames
Although we don't yet emit or receive SID frames, this will make it
possible to deal with them when the time comes.

At this point, everything should be ready on the decoder side; SID
frames (those within a G.729 stream, not those sent via RFC 3389) will
be decoded as expected, and untransmitted frames will be treated as
erased frames, where `conceal()` will end up generating the comfort
noise.

On the encoder side, as indicated by the comments, two issues remain:
one regarding SID frames (if they do not happen to be at the end of the
payload), and one with an empty payload of only untransmitted frames (if
it ends up being transmitted nonetheless as a NAT keepalive).