Reconstruct the parameters string in t_sdp::get_fmtp()
t_sdp::get_fmtp() previously assumed that any "a=fmtp:" line contained
only two tokens, but RFC 4566 does not preclude the presence of spaces
in the "format specific parameters" part.  In fact, there are many
examples (such as found in RFC 6871) where multiple parameters are
separated by "; ", thus creating several tokens.

The proper fix would probably be to introduce a `maxsplit` argument to
split_ws() in order to keep the whole string intact, but that would be a
bit intrusive.  Simply putting some spaces back to reconstruct the
string to a somewhat equivalent version will be enough for our needs.