Add -DDEBUG to non-release builds, to prevent uCommon from adding NDEBUG
Including <ucommon/ucommon.h> will (as of uCommon 7.0.0) unilaterally
define NDEBUG, and thus compile away any assert(), unless DEBUG has been
previously defined.[1]  Until this is fixed, the easiest way around this
is simply to automatically define DEBUG for non-release builds.

Note that there are a few other header files that check for the presence
of DEBUG[2], but the net effect appears to be non-significant.  (It *is*
a debug build, after all!)

 [1] https://lists.gnu.org/archive/html/bug-commoncpp/2019-12/msg00000.html

 [2] - libxml2/libxml/xmlmemory.h
     - X11/Xthreads.h
     - X11/extensions/lbxproto.h
     - FLAC/assert.h