Switch Travis CI distribution to bionic
Older versions of ccrtp under trusty/xenial do not compile with g++ 7+,
so moving on to bionic will finally allow us to test modern compilers.
(Ironically, current ccrtp will no longer compile with g++ 4.9.)

Note that libzrtpcpp is no longer available on bionic, so it was
disabled from the tests.