Relax GetProfileNameForm validator to only check individual characters
QRegExpValidator can be somewhat user-unfriendly, as it will prevent
characters from being input, or even deleted, while leaving the user in
the dark as to what is going on.

Retricting the set of characters allowed shouldn't pose a problem if we
make a mention of this on the form.  Checking for a leading "." or "@",
however, is better left at submission time, with an explicit error
message in that case.