CI: Unfold the bcg729 download/build one-liner
This is so much more readable. ☺

Note that GitHub Actions invokes the shell (either `bash` or `sh`) with
the `-e` option, so all those `&&` are no longer necessary.  Also, each
job runs in its own shell, so isolating `cd` is no longer required.

We also take this opportunity to rename the BUILD_BCG729* environment
variables and matrix keys, while also turning the latter into lowercase
as will be done everywhere else in the next commit.