CI: Trim down the list of compiler versions we test
This previously long list is now trimmed down to merely the lowest and
highest versions of each compiler.  This is based on the assumption that
compilation failures will either occur on compiler version "X and above"
(if caused by a change in the compiler's behavior), or "older than Y"
(if caused by our use of a feature unsupported in older versions).

Just to be safe, we should also make sure to test at least one version
on each Ubuntu release (which is already the case at the moment).