Add missing `$$` initializations in `sdp_parser.yxx`
These two rules used `$$` without ever initializing it, which is Not
Good™.  They happened to work nonetheless due to the fact that `yyval`
still contained the value returned by `$1` at that point, but that
should definitely not be relied on.

This also gets rid of the `unused value: $1 [-Wother]` warnings.