uic: Remove obsolete `<images>` element from `messageform.ui`
The `<images>` element was a Qt 3 feature¹, now unsupported by Qt 5.
While it could certainly be converted to something modern, its only
purpose here is to provide a `<pixmap>` to our `<customwidget>`, which
is "displayed on a Qt Designer toolbar button which the user can use to
create an instance of the custom widget"².  This is something that we
definitely can live without.  :smile:

(The actual pixmap was nothing more than the Qt 3 logo; I guess it was
added by default by Qt Designer.)

 ¹ https://doc.qt.io/archives/3.3/designer-manual-16.html#1-10
 ² https://doc.qt.io/archives/3.3/designer-manual-16.html#1-6