Solve (almost) all shift/reduce conflicts in `parser.yxx`
The shift/reduce conflicts were due to these rules:

    comment:     %empty | '(' T_COMMENT ')'
    server:      comment
    hdr_server:  server | hdr_server server

Since `comment`, and therefore `server`, can be empty, this allows for a
server list to be parsed in an infinite number of ways, with a bunch of
empty strings thrown in.

The solution is to disallow `comment` being empty, and introduce a new
rule (`comment_opt`) to represent the old behavior.  `comment` is thus
replaced with `comment_opt` everywhere except in the first `server`
rule, which is where the conflict was born.

Unfortunately, due to how our lexer strips whitespace between tokens, it
is still impossible to tell the difference between `Foo(Bar)`¹ and `Foo
(Bar)`, hence the remaining conflict.

 ¹ Technically not allowed by RFC 3261, but in use nonetheless.

This commit was tested against a `Server` header with the string
`FPBX-15.0.17.34(16.17.0) Foo/1.2.3 Bar Baz (comment) (another
comment)`, which was parsed correctly.  A few tests were also performed
with `Retry-After` to make sure that no bugs were introduced.