Add `cmake_minimum_required()` `policy_max` argument to silence warnings
This basically lets CMake know that, while we don't require any version
higher than `2.8.10`, we are still compatible with all polices present
in `3.22`.

(Although support for this argument was added in 3.12, older CMake
versions simply ignore it instead of breaking.)

We also take the opportunity to move the call above `project()`, as
recommended by the documentation.