Display any provided Reason header when receiving a BYE or CANCEL
RFC 3326 defines a `Reason` header field which can be added to a request
in order to provide some explanation for why that request was issued.
It is mostly useful for `BYE` and `CANCEL` requests, so that the UA can
explain why a call or `INVITE` was terminated.

While we don't bother generating such headers (yet), we should at least
be kind enough to display those which we receive.