Skip C++11 regular expressions check when cross-compiling
The C++11 regular expressions check added in #117 involves compiling and
running a test program.  When cross-compiling, this is no longer
possible, and calling check_cxx_source_runs() will cause the build to
fail.

Since libstdc++ 4.8 is pretty much obsolete by now and can only be found
in old enterprise LTS releases¹ that are near their end-of-life, we may
as well skip that check when cross-compiling and simply assume that
everything is working fine.

 ¹ Distribution releases still using libstdc++ 4.8 as of today:
     - Ubuntu 14.04 LTS (Trusty Tahr)
     - Red Hat Enterprise Linux 7
     - SUSE Linux Enterprise Server 12