qt5+: Pass single `StandardButtons` arg. to `QMessageBox` funcs [4/4]
Invoking `QMessageBox` functions with separate `button0`, `button1`, ...
arguments was deprecated in Qt 4; all button specifications are now
passed within a single `QMessageBox::StandardButtons` argument.

This commit takes care of the last such instance, where buttons were
passed this time as `QString`.  This requires calling `addButton()` for
each one, as well as replacing the following `switch` with an `if`,
since the return type of `clickedButton()` is a (non-constant) pointer.
(For the record, since we are using custom buttons, the return value of
`exec()` is now an opaque value.)