Prevent our own `gsm.h` from being shadowed
`libgsm` is typically installed with a top-level `gsm.h` include file
(either as the main file or as a compatibility symlink to `gsm/gsm.h`).
If `libgsm` is installed but `WITH_GSM` is disabled, it's possible that
this copy will shadow our own when `audio/gsm/src` is being built,
depending on how include directories are ordered.

This will typically cause a build failure on systems such as FreeBSD,
where `/usr/local/include` is inserted early in the list, such as seen
in https://github.com/LubosD/twinkle/pull/305#issuecomment-1285847851.

(While both copies are almost identical, the original one enforces C
linkage, which ours does not.)